feat: entity 어노테이션 추가, id 타입을 Int로 변경

+++ app/src/main/java/com/hongul/fliq/data/card/CardEntity.kt
@@ -4,8 +4,8 @@ import androidx.room.Entity
 
 @Entity(tableName = "cards")
 data class CardEntity(
-    val id: Long,
-    val owner: Long,
+    val id: Int,
+    val owner: Int,
     val name: String,
     val phone: String,
     val email: String,
